Assignment2: added report placeholder
Added the placeholder for the report, as well as the Makefile targets for
turning the LaTeX into a PDF for submission. It still needs a test script to help
generate a graph for the report.

+++ Assignments/Assignment2/include/page.h
@@ -17,7 +17,7 @@ struct s_page
 {
     int number;
     int valid;
-    int accessed;
+    unsigned long accessed;
 };
 
 typedef struct s_page page;
